Pass blocks using redux connect
